+++ src/ast/import_visit.rs
@@ -70,6 +70,7 @@ impl ImportTargetValidator<'_> {
             PackageBodyElement::VerificationCaseUsage(n) => self.use_case_body(&n.value.body),
             PackageBodyElement::ViewUsage(n) => self.view_body(&n.value.body),
             PackageBodyElement::Error(_)
+            | PackageBodyElement::Unsupported(_)
             | PackageBodyElement::Doc(_)
             | PackageBodyElement::Comment(_)
             | PackageBodyElement::TextualRep(_)